pyproject.toml
[project]
name = "databricks-multi-tenant-billing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.14"
dependencies = [
    "databricks-sdk>=0.73.0",
    "databricks-vectorsearch>=0.63",
    "pydantic-settings>=2.12.0",
    "pyspark>=4.0.1",
    "streamlit>=1.51.0",
]
